refactor: move harmonization and staging logic from compute_equity_stake.py to clean_and_prepare_raw_data.py
